add check file and dir

+++ outFile.js
@@ -1,6 +1,6 @@
 const fs = require('fs')
 
-module.exports = async (srcBaseDir = `${__dirname}/node_modules/skybase-stat/`) => {
+module.exports = async (destBaseDir = '.', srcBaseDir = `${__dirname}/node_modules/skybase-stat/`) => {
   let obj = {
     'router': {
       'skyapi': {
@@ -34,7 +34,13 @@ module.exports = async (srcBaseDir = `${__dirname}/node_modules/skybase-stat/`) 
     }
   }
   try {
-    await outPutFile('.', '', obj, srcBaseDir)
+    let isExist = fs.existsSync(srcBaseDir) // 判断目录是否存在
+    if (!isExist) {
+      console.error(srcBaseDir, '目录不存在！')
+      return
+    }
+
+    await outPutFile(destBaseDir, '', obj, srcBaseDir)
   } catch (e) {
     console.error(e.stack)
   }
@@ -45,11 +51,23 @@ async function outPutFile(dir, key, obj, srcBaseDir) {
     // console.log(`创建目录  ${dir}`)
     await fs.mkdirSync(dir, { recursive: true })
     // console.log(`原：${srcBaseDir}${obj}  目的：${dir}/${key}`)
+    if (!await checkFileExist(`${srcBaseDir}${obj}`)) {
+      console.error(`原文件不存在\t${srcBaseDir}${obj}`)
+      return
+    }
     await fs.copyFileSync(`${srcBaseDir}${obj}`, `${dir}/${key}`);
     return
   }
   for (let k in obj) {
-    outPutFile(key === '' ? '.' : `${dir}/${key}`, k, obj[k], srcBaseDir)
+    outPutFile(key === '' ? dir : `${dir}/${key}`, k, obj[k], srcBaseDir)
   }
+}
+
+async function checkFileExist(filePath) {
+  try {
+    await fs.accessSync(filePath, fs.constants.R_OK)
+    return true
+  } catch (e) { }
+  return false
 }
